requirements.txt
djangorestframework
django
django-mailgun
django-cors-headers